Make sure that conditional expressions are optimized correctly
The constant directory keeps a map from scoped-names to their value nodes,
so when optimizing the value we have to do so in place (this isn't
a great design, and I'll fix it in a later commit). Without this change
a constant that was initialized with a constant conditional expression
would have its initializer optimized, but code using that constant
wouldn't have the value substituted.

ie

  const BAR = false;
  const FOO = BAR ? 42 : 24;
  function foo() { return FOO; }

would end up as

  const FOO = 24;
  function foo() { return FOO; }

rather than

  function foo() { return 24; }