Fix caching vs Promise.all
We were caching reads of resource files individually, and groups
of resource files (per target). But it didn't work, because
under Promise.all, each cache entry was checked, and then we
generate it asynchronously, and then fill in the cache entry.

To avoid that, keep Promises in the cache, and have every thread
wait on the promise.

For samuelmr-garmin-abouttime, this goes from 1500 resource files
being parsed, to just 19 (with no duplicates). The time for
get_jungles drops from ~200s to 3s.