Implement globSome to determine if glob would find any files
While parsing jungle files, we run glob to determine whether
particular paths (in sourcePath, resourcePath etc) exist.

Globbing **/*.mc in the root of the project (the default
sourcePath) can be quite slow, so globSome stops after finding
the first file.

Trying to do this with the "glob" module resulted in crashes (it
doesn't seem to handle "abort" well), so I switched to fast-glob,
which also got a small speed increase by itself.