Rewrite control flow to make it more accurate
The original control flow analysis took a naive approach to
handling sub-expression control flow. It essentially assumed
that the control flow in

  if (x && y && z) {
  }

was the same as if it had been written

  var t1 = x && y;
  var t2 = t1 && z;
  if (t2) {
  }

This is functionally correct, and up until now was perfectly good
for PRE (which was the only thing using control-flow).

But type-flow wants a more accurate flow graph, and in particular
wants to be able to track conditions like

  if (x != null && x.y != null) {
    x.y.foo();
  }

which simply doesn't work with the original formulation.