Fix a bug finding the next component in a MemberExpression
Given something like $.Toybox.Application.AppBase.Properties,
when we lookup Properties in AppBase we don't find it.
So we have to check in Application and Toybox and finally $,
but we weren't doing that.

Note that Garmin's type checker throws an error for such cases,
so its not normally an issue. But untypechecked code might do
this and we shouldn't throw an error (or worse, make an incorrect
optimization decision) as a result of not finding the correct
target.