Fix some issues with constant folding and inlining
When inlining, the inliner didn't update the type-stack
result of the function - so even when it returned a literal,
the containing expression couldn't get folded.

But even for non-literals, we could miss some optimization
opportunities, so re-optimize VariableDeclarations if
any inlining took place (just as we do for function bodies).