Attach line number info to bytecodes
This lets us fully recompute the line number tables (both
the embedded one, and the one in the debug.xml) rather
than simply trying to update the existing entries by
remapping the pc.

It also means we don't have to break blocks at every pc
in the line number table, making for bigger basic blocks,
and better (for now) optimizations.