Fix "Minimize Modules" in background/glance scopes
It seems that $.Rez is dynamically bound to globals/BackgroundRez when
running in background scope, and to globals/GlanceRez when running in
glance scope.

But `using Rez;` unconditionally binds it to globals/Rez causing
background and glance code that tries to use it to fail.

As a result, we can only use `using Rez;` in code that's not annotated
`:background` or `:glance`