postbuild: Don't insert a pre-definition before an frpush
Pre definitions are normally inserted ahead of the first use (if there is one)
in the dominating block. But we also ensure that they're inserted prior to the
first potentially throwing bytecode if we're inside a try-catch. There was a
bug where we would insert 1 ahead of the invokem; if the call had no arguments,
this would be before the frpush, which is stateful, and potentially affected by
an inserted getm. This ensures that we insert directly before the invokem.