Fix bytecode optimizer to handle shrv/shlv without a following nop
Earlier sdks always treated shrv and shlv as a two byte opcode with
a following nop. Although the simulator didn't require the nop,
removing it resulted in the Ciq Store reporting that the binary
was corrupt - so we had to continue to treat them as two byte opcodes.

In CIQ 5, the compiler correctly generates one byte opcodes, so we
have to parse them as such. On output, we check the ciq-version and
insert nops if necessary.